move fido2 popup components from vault to autofill (#10381)
