+++ apps/browser/src/autofill/popup/fido2/fido2-cipher-row.component.html
